fix: overflow

+++ src/public/administration/write/OnlyEmergencyDeleverager.sol
@@ -39,7 +39,7 @@ abstract contract OnlyEmergencyDeleverager is
         nonReentrant
     {
         require(
-            deleverageFeeDividend * maxDeleverageFeeDivider <= deleverageFeeDivider * maxDeleverageFeeDividend,
+            uint256(deleverageFeeDividend) * maxDeleverageFeeDivider <= uint256(deleverageFeeDivider) * maxDeleverageFeeDividend,
             ExceedsMaxDeleverageFee(
                 deleverageFeeDividend, deleverageFeeDivider, maxDeleverageFeeDividend, maxDeleverageFeeDivider
             )
